requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
backcall==0.2.0
bcrypt==4.0.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleanco==2.2
click==8.1.7
comm==0.1.4
connectorx==0.3.2
cryptography==41.0.3
Cython==0.29.36
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
elastic-transport==8.4.0
elasticsearch==8.4.2
exceptiongroup==1.1.3
executing==1.2.0
frozenlist==1.4.0
hdbscan==0.8.33
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
llvmlite==0.41.0
loguru==0.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numba==0.58.0
numpy==1.25.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
pandasticsearch==0.6.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pymongo==4.5.0
PyNaCl==1.5.0
pynndescent==0.5.10
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
sb-py-fw @ git+ssh://git@github.com/startup-bakery/sb-py-fw@4444c2dabf98d4136582dc9a2b8ccdb8bb89613b
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.20
sshtunnel==0.4.0
stack-data==0.6.2
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
umap==0.1.1
umap-learn==0.5.4
urllib3==1.26.16
validators==0.22.0
wcwidth==0.2.6
weaviate-client==3.24.1
yarl==1.9.2
zipp==3.16.2
